Introduce a service/use-case layer to keep FastAPI routers thin and centralize business logic across the API (adaptation, tasks, focus, interactions, auth, users).
