+++ app/api/auth.py
@@ -1,70 +1,42 @@
-from fastapi import APIRouter, Depends, HTTPException, status
-from sqlalchemy import select
+from fastapi import APIRouter, Depends
+from sqlalchemy.ext.asyncio import AsyncSession
 
-from app.core.security import verify_password, get_password_hash, create_access_token, create_refresh_token, decode_token
+from app.api.deps import get_current_user
 from app.db.session import get_db
 from app.models.user import User
-from app.schemas.user import UserCreate, UserResponse, Token
-from app.api.deps import get_current_user
-from sqlalchemy.ext.asyncio import AsyncSession
+from app.schemas.auth import LoginRequest, RefreshBody
+from app.schemas.user import Token, UserCreate, UserResponse
+from app.services.auth_service import (
+    login as login_service,
+)
+from app.services.auth_service import (
+    me as me_service,
+)
+from app.services.auth_service import (
+    refresh as refresh_service,
+)
+from app.services.auth_service import (
+    register as register_service,
+)
 
 router = APIRouter()
 
 
 @router.post("/register", response_model=UserResponse)
 async def register(body: UserCreate, db: AsyncSession = Depends(get_db)):
-    result = await db.execute(select(User).where(User.email == body.email))
-    if result.scalar_one_or_none():
-        raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail="Email already registered")
-    user = User(
-        email=body.email,
-        name=body.name,
-        password_hash=get_password_hash(body.password),
-    )
-    db.add(user)
-    await db.commit()
-    await db.refresh(user)
-    return user
-
-
-from pydantic import BaseModel
-
-
-class LoginRequest(BaseModel):
-    email: str
-    password: str
+    return await register_service(db=db, body=body)
 
 
 @router.post("/login", response_model=Token)
 async def login(body: LoginRequest, db: AsyncSession = Depends(get_db)):
-    result = await db.execute(select(User).where(User.email == body.email))
-    user = result.scalar_one_or_none()
-    if not user or not verify_password(body.password, user.password_hash):
-        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Invalid email or password")
-    return Token(
-        access_token=create_access_token(user.id),
-        refresh_token=create_refresh_token(user.id),
-    )
-
-
-class RefreshBody(BaseModel):
-    refresh_token: str
+    return await login_service(db=db, body=body)
 
 
 @router.post("/refresh", response_model=Token)
 async def refresh(body: RefreshBody):
-    payload = decode_token(body.refresh_token)
-    if not payload or payload.get("type") != "refresh":
-        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Invalid refresh token")
-    sub = payload.get("sub")
-    if not sub:
-        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Invalid refresh token")
-    return Token(
-        access_token=create_access_token(sub),
-        refresh_token=create_refresh_token(sub),
-    )
+    return await refresh_service(body=body)
 
 
 @router.get("/me", response_model=UserResponse)
 async def me(current_user: User = Depends(get_current_user)):
-    return current_user
+    return await me_service(current_user=current_user)
