+++ app/api/focus.py
@@ -1,14 +1,27 @@
 from uuid import UUID
-from datetime import datetime, timezone
+
 from fastapi import APIRouter, Depends, HTTPException, Query
-from sqlalchemy import select, desc
 from sqlalchemy.ext.asyncio import AsyncSession
 
 from app.api.deps import get_current_user
 from app.db.session import get_db
 from app.models.user import User
-from app.models.focus_session import FocusSession
-from app.schemas.focus import FocusSessionCreate, FocusSessionUpdate, FocusSessionResponse
+from app.schemas.focus import FocusSessionCreate, FocusSessionResponse, FocusSessionUpdate
+from app.services.focus_service import (
+    end_focus_session as end_focus_session_service,
+)
+from app.services.focus_service import (
+    get_focus_session as get_focus_session_service,
+)
+from app.services.focus_service import (
+    get_my_focus_sessions as get_my_focus_sessions_service,
+)
+from app.services.focus_service import (
+    list_focus_sessions as list_focus_sessions_service,
+)
+from app.services.focus_service import (
+    start_focus_session as start_focus_session_service,
+)
 
 router = APIRouter()
 
@@ -19,16 +32,7 @@ async def start_focus_session(
     current_user: User = Depends(get_current_user),
     db: AsyncSession = Depends(get_db),
 ):
-    now = datetime.now(timezone.utc)
-    session = FocusSession(
-        user_id=current_user.id,
-        task_id=body.task_id,
-        start_time=now,
-    )
-    db.add(session)
-    await db.commit()
-    await db.refresh(session)
-    return session
+    return await start_focus_session_service(db=db, current_user=current_user, body=body)
 
 
 @router.get("", response_model=list[FocusSessionResponse])
@@ -37,14 +41,7 @@ async def list_focus_sessions(
     current_user: User = Depends(get_current_user),
     db: AsyncSession = Depends(get_db),
 ):
-    stmt = (
-        select(FocusSession)
-        .where(FocusSession.user_id == current_user.id)
-        .order_by(desc(FocusSession.start_time))
-        .limit(limit)
-    )
-    result = await db.execute(stmt)
-    return list(result.scalars().all())
+    return await list_focus_sessions_service(db=db, current_user=current_user, limit=limit)
 
 
 @router.get("/sessions", response_model=list[FocusSessionResponse])
@@ -53,7 +50,7 @@ async def get_my_focus_sessions(
     current_user: User = Depends(get_current_user),
     db: AsyncSession = Depends(get_db),
 ):
-    return await current_user.get_focus_sessions(db, limit=limit)
+    return await get_my_focus_sessions_service(db=db, current_user=current_user, limit=limit)
 
 
 @router.get("/{session_id}", response_model=FocusSessionResponse)
@@ -62,14 +59,8 @@ async def get_focus_session(
     current_user: User = Depends(get_current_user),
     db: AsyncSession = Depends(get_db),
 ):
-    result = await db.execute(
-        select(FocusSession).where(
-            FocusSession.id == session_id,
-            FocusSession.user_id == current_user.id,
-        )
-    )
-    session = result.scalar_one_or_none()
-    if not session:
+    session = await get_focus_session_service(db=db, current_user=current_user, session_id=session_id)
+    if not session:  # Preserve previous HTTP semantics
         raise HTTPException(status_code=404, detail="Focus session not found")
     return session
 
@@ -81,21 +72,7 @@ async def end_focus_session(
     current_user: User = Depends(get_current_user),
     db: AsyncSession = Depends(get_db),
 ):
-    result = await db.execute(
-        select(FocusSession).where(
-            FocusSession.id == session_id,
-            FocusSession.user_id == current_user.id,
-        )
-    )
-    session = result.scalar_one_or_none()
-    if not session:
+    session = await end_focus_session_service(db=db, current_user=current_user, session_id=session_id, body=body)
+    if not session:  # Preserve previous HTTP semantics
         raise HTTPException(status_code=404, detail="Focus session not found")
-    if body.end_time is not None or body.completed is not None:
-        session.end_session(body.end_time)
-    if body.interruptions is not None:
-        session.interruptions = body.interruptions
-    if body.notes is not None:
-        session.notes = body.notes
-    await db.commit()
-    await db.refresh(session)
     return session
