+++ app/api/tasks.py
@@ -1,13 +1,30 @@
 from uuid import UUID
-from fastapi import APIRouter, Depends, HTTPException, status, Query
-from sqlalchemy import select
+
+from fastapi import APIRouter, Depends, HTTPException, Query, status
 from sqlalchemy.ext.asyncio import AsyncSession
 
 from app.api.deps import get_current_user
 from app.db.session import get_db
 from app.models.user import User
-from app.models.task import Task, TaskStatus
-from app.schemas.task import TaskCreate, TaskUpdate, TaskResponse
+from app.schemas.task import TaskCreate, TaskResponse, TaskUpdate
+from app.services.tasks_service import (
+    create_task as create_task_service,
+)
+from app.services.tasks_service import (
+    delete_task as delete_task_service,
+)
+from app.services.tasks_service import (
+    get_active_tasks as get_active_tasks_service,
+)
+from app.services.tasks_service import (
+    get_task as get_task_service,
+)
+from app.services.tasks_service import (
+    list_tasks as list_tasks_service,
+)
+from app.services.tasks_service import (
+    update_task as update_task_service,
+)
 
 router = APIRouter()
 
@@ -18,11 +35,7 @@ async def list_tasks(
     current_user: User = Depends(get_current_user),
     db: AsyncSession = Depends(get_db),
 ):
-    stmt = select(Task).where(Task.user_id == current_user.id).order_by(Task.order_index, Task.deadline)
-    if status_filter:
-        stmt = stmt.where(Task.status == status_filter)
-    result = await db.execute(stmt)
-    return list(result.scalars().all())
+    return await list_tasks_service(db=db, current_user=current_user, status_filter=status_filter)
 
 
 @router.post("", response_model=TaskResponse, status_code=status.HTTP_201_CREATED)
@@ -31,21 +44,7 @@ async def create_task(
     current_user: User = Depends(get_current_user),
     db: AsyncSession = Depends(get_db),
 ):
-    task = Task(
-        user_id=current_user.id,
-        title=body.title,
-        description=body.description,
-        deadline=body.deadline,
-        category=body.category,
-        estimated_effort=body.estimated_effort,
-        order_index=body.order_index,
-        tags=body.tags or [],
-    )
-    task.compute_priority_score()
-    db.add(task)
-    await db.commit()
-    await db.refresh(task)
-    return task
+    return await create_task_service(db=db, current_user=current_user, body=body)
 
 
 @router.get("/active", response_model=list[TaskResponse])
@@ -53,7 +52,7 @@ async def get_active_tasks(
     current_user: User = Depends(get_current_user),
     db: AsyncSession = Depends(get_db),
 ):
-    return await current_user.get_active_tasks(db)
+    return await get_active_tasks_service(db=db, current_user=current_user)
 
 
 @router.get("/{task_id}", response_model=TaskResponse)
@@ -62,9 +61,8 @@ async def get_task(
     current_user: User = Depends(get_current_user),
     db: AsyncSession = Depends(get_db),
 ):
-    result = await db.execute(select(Task).where(Task.id == task_id, Task.user_id == current_user.id))
-    task = result.scalar_one_or_none()
-    if not task:
+    task = await get_task_service(db=db, current_user=current_user, task_id=task_id)
+    if not task:  # Preserve previous HTTP semantics
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Task not found")
     return task
 
@@ -76,31 +74,9 @@ async def update_task(
     current_user: User = Depends(get_current_user),
     db: AsyncSession = Depends(get_db),
 ):
-    result = await db.execute(select(Task).where(Task.id == task_id, Task.user_id == current_user.id))
-    task = result.scalar_one_or_none()
-    if not task:
+    task = await update_task_service(db=db, current_user=current_user, task_id=task_id, body=body)
+    if not task:  # Preserve previous HTTP semantics
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Task not found")
-    if body.title is not None:
-        task.title = body.title
-    if body.description is not None:
-        task.description = body.description
-    if body.deadline is not None:
-        task.deadline = body.deadline
-    if body.category is not None:
-        task.category = body.category
-    if body.status is not None:
-        task.update_status(body.status)
-    if body.estimated_effort is not None:
-        task.estimated_effort = body.estimated_effort
-    if body.actual_time_spent is not None:
-        task.actual_time_spent = body.actual_time_spent
-    if body.order_index is not None:
-        task.order_index = body.order_index
-    if body.tags is not None:
-        task.tags = body.tags
-    task.compute_priority_score()
-    await db.commit()
-    await db.refresh(task)
     return task
 
 
@@ -110,9 +86,7 @@ async def delete_task(
     current_user: User = Depends(get_current_user),
     db: AsyncSession = Depends(get_db),
 ):
-    result = await db.execute(select(Task).where(Task.id == task_id, Task.user_id == current_user.id))
-    task = result.scalar_one_or_none()
-    if not task:
+    deleted = await delete_task_service(db=db, current_user=current_user, task_id=task_id)
+    if not deleted:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Task not found")
-    await db.delete(task)
-    await db.commit()
+    return None
