+++ app/api/users.py
@@ -1,17 +1,19 @@
-from fastapi import APIRouter, Depends, HTTPException, status
+from fastapi import APIRouter, Depends
 from sqlalchemy.ext.asyncio import AsyncSession
 
 from app.api.deps import get_current_user
 from app.db.session import get_db
 from app.models.user import User
-from app.schemas.user import UserUpdate, UserResponse
+from app.schemas.user import UserResponse, UserUpdate
+from app.services.users_service import get_me as get_me_service
+from app.services.users_service import update_me as update_me_service
 
 router = APIRouter()
 
 
 @router.get("/me", response_model=UserResponse)
 async def get_me(current_user: User = Depends(get_current_user)):
-    return current_user
+    return await get_me_service(current_user=current_user)
 
 
 @router.patch("/me", response_model=UserResponse)
@@ -20,10 +22,4 @@ async def update_me(
     current_user: User = Depends(get_current_user),
     db: AsyncSession = Depends(get_db),
 ):
-    if body.name is not None:
-        current_user.name = body.name
-    if body.study_preferences is not None:
-        current_user.study_preferences = body.study_preferences
-    await db.commit()
-    await db.refresh(current_user)
-    return current_user
+    return await update_me_service(db=db, current_user=current_user, body=body)
