+++ app/schemas/auth.py
@@ -1,0 +1,11 @@
+from pydantic import BaseModel, EmailStr
+
+
+class LoginRequest(BaseModel):
+    email: EmailStr
+    password: str
+
+
+class RefreshBody(BaseModel):
+    refresh_token: str
+
