+++ app/services/__init__.py
@@ -1,0 +1,5 @@
+"""Application services (use-cases).
+
+Routers should stay thin; services contain orchestration/business logic.
+"""
+
